Refactor filename parsing code to deal with non addr-Y-M-D.xxx filenames
Like if someone put "addr-Y-description-" or something.

TODO probably could be more elegant.

What to do if we get something like "number-32.jpg" -- that'd be interpreted as
address == number
year    == 32

At the moment these are allowed:

Anystring-Y.ext
Anystring-Y-M.ext
Anystring-Y-M-D.ext
Anystring-Y-M-D-whatever....ext

The numericality of Y, M and D are validated, though not the sanity. THis may create later problems with the 32th of January and such.